requirements.txt
Django==2.2
djangorestframework==3.10.2
SQLAlchemy==1.2.15
sqlparse==0.3.0
python-decouple==3.1
psycopg2==2.8.3
gunicorn==19.9.0
dj-database-url==0.5.0
dj-static==0.0.6
whitenoise==3.2